Test reading and processing JBang metadata blocks (#20)
We make JythonCli testable with JUnit, by changing readJBangBlock to accept a Reader. Then we give it test strings that express mistakes a programmer might make in the jbang block. We expect corresponding kinds of failure. We add a short entry to the README about using it.

Failing tests are disabled until addressed by parser change.